src/5. Tools and Agents/3_multi_agent_systems/3_Google_A2A_protocol/standalone agents/A2A_standalone_agent_math/agent_tests.py
from agent import Agent
from dotenv import load_dotenv
load_dotenv()

agent = Agent()
sessionID = "session_12345"

query = "How much is 100 / 4?"
response = agent.invoke(query, sessionID)
print(response)